requirements.txt
streamlit
easyocr
opencv-python
pillow
pandas
lxml